backend/requirements.txt
Flask
PyJWT
flask-cors
pymongo
python-dotenv
cryptography>=41.0.0
pytest>=7.0.0
bcrypt>=4.0.0
flask-limiter>=3.0.0
gunicorn>=21.0.0
webauthn>=2.0.0
